# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanLedger API - Loan management with interest and commission tracking
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface User {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  isActive?: boolean;
  createdAt: string;
  updatedAt?: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserInput {
  email: string;
  name: string;
  phone?: string;
  role: UserInputRole;
}

export interface UserUpdate {
  name?: string;
  phone?: string;
}

export type UserAdminUpdateRole = typeof UserAdminUpdateRole[keyof typeof UserAdminUpdateRole];


export const UserAdminUpdateRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserAdminUpdate {
  name?: string;
  phone?: string;
  role?: UserAdminUpdateRole;
  isActive?: boolean;
}

export type BorrowerStatus = typeof BorrowerStatus[keyof typeof BorrowerStatus];


export const BorrowerStatus = {
  active: 'active',
  closed: 'closed',
  defaulted: 'defaulted',
} as const;

export interface Borrower {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  name: string;
  address: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** Loan principal amount in currency units */
  principalAmount: number;
  /** Total monthly interest rate as percentage (e.g. 15 = 15%) */
  interestRate: number;
  /** Base interest rate, fixed at 10% */
  baseInterestRate: number;
  /** Commission = interestRate - baseInterestRate (e.g. 5% if rate=15%) */
  commissionRate: number;
  /**
     * Loan tenure in months (optional)
     * @nullable
     */
  tenure?: number | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  status: BorrowerStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  totalInterestEarned?: number | null;
  /** @nullable */
  totalCommissionEarned?: number | null;
  /** @nullable */
  monthsElapsed?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export interface BorrowerInput {
  name: string;
  address: string;
  phone?: string;
  email?: string;
  principalAmount: number;
  /**
     * Total monthly interest % (default 10, can be higher)
     * @minimum 0
     */
  interestRate: number;
  /** Tenure in months (optional) */
  tenure?: number;
  startDate: string;
  endDate?: string;
  notes?: string;
}

export type BorrowerUpdateStatus = typeof BorrowerUpdateStatus[keyof typeof BorrowerUpdateStatus];


export const BorrowerUpdateStatus = {
  active: 'active',
  closed: 'closed',
  defaulted: 'defaulted',
} as const;

export interface BorrowerUpdate {
  name?: string;
  address?: string;
  phone?: string;
  email?: string;
  tenure?: number;
  endDate?: string;
  status?: BorrowerUpdateStatus;
  notes?: string;
}

export interface Payment {
  id: number;
  borrowerId: number;
  /**
     * @minimum 1
     * @maximum 12
     */
  month: number;
  year: number;
  principalAmount: number;
  interestRate: number;
  baseInterestRate: number;
  commissionRate: number;
  /** Total interest amount (principalAmount * interestRate / 100) */
  interestAmount: number;
  /** Base interest portion (principalAmount * 10 / 100) */
  baseInterestAmount: number;
  /** Commission portion (principalAmount * commissionRate / 100) */
  commissionAmount: number;
  isPaid: boolean;
  /** @nullable */
  paidDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PaymentInput {
  /**
     * @minimum 1
     * @maximum 12
     */
  month: number;
  year: number;
  isPaid?: boolean;
  paidDate?: string;
  notes?: string;
}

export interface PaymentUpdate {
  isPaid?: boolean;
  paidDate?: string;
  notes?: string;
}

export interface DashboardStats {
  totalBorrowers: number;
  activeBorrowers: number;
  closedBorrowers?: number;
  defaultedBorrowers?: number;
  totalPrincipal: number;
  totalInterestEarned: number;
  totalCommissionEarned: number;
  totalBaseInterestEarned: number;
  pendingPaymentsCount: number;
  currentMonthInterest: number;
  currentMonthCommission: number;
  currentMonthBaseInterest?: number;
}

export interface MonthlyStats {
  month: number;
  year: number;
  totalInterest: number;
  baseInterest: number;
  commission: number;
  paymentsCount: number;
  paidCount: number;
  /** @nullable */
  principalOutstanding?: number | null;
}

export interface YearlyStats {
  year: number;
  totalInterest: number;
  baseInterest: number;
  commission: number;
  paymentsCount: number;
  paidCount: number;
}

export interface UserStats {
  userId: number;
  userName: string;
  userEmail?: string;
  totalBorrowers: number;
  activeBorrowers?: number;
  totalPrincipal: number;
  totalInterest: number;
  totalBaseInterest: number;
  totalCommission: number;
}

export interface AdminDashboard {
  totalUsers: number;
  activeUsers: number;
  totalBorrowers: number;
  activeBorrowers: number;
  totalPrincipalDeployed: number;
  totalInterestEarned: number;
  totalBaseInterestEarned: number;
  totalCommissionEarned: number;
  monthlyTrend: MonthlyStats[];
  topPerformers: UserStats[];
}

export type ListBorrowersParams = {
status?: ListBorrowersStatus;
search?: string;
};

export type ListBorrowersStatus = typeof ListBorrowersStatus[keyof typeof ListBorrowersStatus];


export const ListBorrowersStatus = {
  active: 'active',
  closed: 'closed',
  defaulted: 'defaulted',
} as const;

export type GetMonthlyStatsParams = {
year?: number;
/**
 * @nullable
 */
userId?: number | null;
};

export type GetYearlyStatsParams = {
/**
 * @nullable
 */
userId?: number | null;
};