# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LoanLedger API - Loan management with interest and commission tracking
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /borrowers:
    get:
      operationId: listBorrowers
      tags: [borrowers]
      summary: List borrowers (own for user, all for admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, closed, defaulted]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of borrowers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Borrower"
    post:
      operationId: createBorrower
      tags: [borrowers]
      summary: Create a new loan borrower
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowerInput"
      responses:
        "201":
          description: Created borrower
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrower"
  /borrowers/{id}:
    get:
      operationId: getBorrower
      tags: [borrowers]
      summary: Get borrower details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Borrower details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrower"
        "404":
          description: Not found
    patch:
      operationId: updateBorrower
      tags: [borrowers]
      summary: Update borrower basic info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowerUpdate"
      responses:
        "200":
          description: Updated borrower
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrower"
    delete:
      operationId: deleteBorrower
      tags: [borrowers]
      summary: Delete a borrower
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PAYMENTS ────────────────────────────────────────────────────────────────
...
  /analytics/monthly:
    get:
      operationId: getMonthlyStats
      tags: [analytics]
      summary: Monthly interest and commission breakdown
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Monthly stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyStats"
  /analytics/yearly:
    get:
      operationId: getYearlyStats
      tags: [analytics]
      summary: Yearly interest and commission breakdown
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Yearly stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YearlyStats"
...
components:
  schemas:
...
    Borrower:
      type: object
      required: [id, userId, name, address, principalAmount, interestRate, baseInterestRate, commissionRate, startDate, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        name:
          type: string
        address:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        principalAmount:
          type: number
          description: Loan principal amount in currency units
        interestRate:
          type: number
          description: Total monthly interest rate as percentage (e.g. 15 = 15%)
        baseInterestRate:
          type: number
          description: Base interest rate, fixed at 10%
        commissionRate:
          type: number
          description: Commission = interestRate - baseInterestRate (e.g. 5% if rate=15%)
        tenure:
          type: ["integer", "null"]
          description: Loan tenure in months (optional)
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [active, closed, defaulted]
        notes:
          type: ["string", "null"]
        totalInterestEarned:
          type: ["number", "null"]
        totalCommissionEarned:
          type: ["number", "null"]
        monthsElapsed:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BorrowerInput:
      type: object
      required: [name, address, principalAmount, interestRate, startDate]
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        principalAmount:
          type: number
        interestRate:
          type: number
          minimum: 0
          description: Total monthly interest % (default 10, can be higher)
        tenure:
          type: integer
          description: Tenure in months (optional)
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: string
    BorrowerUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        tenure:
          type: integer
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, closed, defaulted]
        notes:
          type: string

    # ── Payment ────────────────────────────────────────────────────────────────
...
    MonthlyStats:
      type: object
      required: [month, year, totalInterest, baseInterest, commission, paymentsCount, paidCount]
      properties:
        month:
          type: integer
        year:
          type: integer
        totalInterest:
          type: number
        baseInterest:
          type: number
        commission:
          type: number
        paymentsCount:
          type: integer
        paidCount:
          type: integer
        principalOutstanding:
          type: ["number", "null"]
    YearlyStats:
      type: object
      required: [year, totalInterest, baseInterest, commission, paymentsCount, paidCount]
      properties:
        year:
          type: integer
        totalInterest:
          type: number
        baseInterest:
          type: number
        commission:
          type: number
        paymentsCount:
          type: integer
        paidCount:
          type: integer
...