# lib/api-zod/src/generated/types/borrowerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanLedger API - Loan management with interest and commission tracking
 * OpenAPI spec version: 0.1.0
 */

export interface BorrowerInput {
  name: string;
  address: string;
  phone?: string;
  email?: string;
  principalAmount: number;
  /**
     * Total monthly interest % (default 10, can be higher)
     * @minimum 0
     */
  interestRate: number;
  /** Tenure in months (optional) */
  tenure?: number;
  startDate: Date;
  endDate?: Date;
  notes?: string;
}