# lib/api-zod/src/generated/types/borrowerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanLedger API - Loan management with interest and commission tracking
 * OpenAPI spec version: 0.1.0
 */
import type { BorrowerUpdateStatus } from './borrowerUpdateStatus';

export interface BorrowerUpdate {
  name?: string;
  address?: string;
  phone?: string;
  email?: string;
  tenure?: number;
  endDate?: Date;
  status?: BorrowerUpdateStatus;
  notes?: string;
}